create apps, add Docker and Docker compose
