+++ .envexample
@@ -4,3 +4,4 @@ DB_PASSWORD=
 DB_HOST=
 DB_PORT=
 SECRET_KEY=
+ALLOWED_HOSTS=
