+++ carmarket/car/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+
+from car.models import Car
+
+class CarSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Car
+        fields = '__all__'
