+++ carmarket/car/views.py
@@ -1,0 +1,11 @@
+from rest_framework import viewsets
+
+from car.models import  Car
+from car.serializers import CarSerializer
+
+
+class CarViewSet(viewsets.ModelViewSet):
+    queryset = Car.objects.all()
+    serializer_class = CarSerializer
+
+
