+++ carmarket/carmarket/settings.py
@@ -28,13 +28,11 @@ SECRET_KEY = os.getenv('SECRET_KEY')
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
 
-ALLOWED_HOSTS = []
-
+ALLOWED_HOSTS = os.getenv('ALLOWED_HOSTS').split(' ')
 
 # Application definition
 
 INSTALLED_APPS = [
-    'main',
     'rest_framework',
     'django_countries',
     'django.contrib.admin',
@@ -43,6 +41,11 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'main',
+    'car',
+    'showroom',
+    'user',
+    'order'
 ]
 
 MIDDLEWARE = [
