--- carmarket/main/serializers.py
@@ -1,38 +1,0 @@
-from rest_framework import serializers
-
-from .models import Buyer, Seller, Showroom, Car, Order, Sale
-
-
-class BuyerSerializers(serializers.ModelSerializer):
-    class Meta:
-        model = Buyer
-        fields = '__all__'
-
-
-class SellerSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Seller
-        fields = "__all__"
-
-class ShowroomSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Showroom
-        fields = "__all__"
-
-
-class CarSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Car
-        fields = '__all__'
-
-
-class OrderSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Order
-        fields = '__all__'
-
-
-class SaleSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Sale
-        fields = '__all__'
