+++ carmarket/main/views.py
@@ -1,40 +1,2 @@
-from rest_framework import viewsets, mixins
-from rest_framework.viewsets import GenericViewSet
-
-from .models import Buyer, Seller, Showroom, Car, Order, Sale
-from .serializers import (BuyerSerializers, SellerSerializer,
-                          ShowroomSerializer, CarSerializer,
-                          OrderSerializer, SaleSerializer)
 
 
-class BuyerViewSet(viewsets.ModelViewSet):
-    queryset = Buyer.objects.all()
-    serializer_class = BuyerSerializers
-
-
-class SellerViewSet(viewsets.ModelViewSet):
-    queryset = Seller.objects.all()
-    serializer_class = SellerSerializer
-
-
-class ShowroomViewSet(viewsets.ModelViewSet):
-    queryset = Showroom.objects.all()
-    serializer_class = ShowroomSerializer
-
-
-class CarViewSet(viewsets.ModelViewSet):
-    queryset = Car.objects.all()
-    serializer_class = CarSerializer
-
-
-class OrderViewSet(mixins.CreateModelMixin,
-                   mixins.RetrieveModelMixin,
-                   mixins.ListModelMixin,
-                   GenericViewSet):
-    queryset = Order.objects.all()
-    serializer_class = OrderSerializer
-
-
-class SaleViewSet(viewsets.ModelViewSet):
-    queryset = Sale.objects.all()
-    serializer_class = SaleSerializer
