+++ carmarket/order/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from order.models import Order
+
+
+class OrderSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Order
+        fields = '__all__'
