+++ carmarket/showroom/serializers.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+
+from showroom.models import Showroom, Sale
+
+
+class ShowroomSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Showroom
+        fields = "__all__"
+
+
+class SaleSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Sale
+        fields = '__all__'
