+++ carmarket/user/views.py
@@ -1,0 +1,15 @@
+from rest_framework import viewsets
+
+from user.models import Buyer, Seller
+from user.serializers import BuyerSerializers, SellerSerializer
+
+
+class BuyerViewSet(viewsets.ModelViewSet):
+    queryset = Buyer.objects.all()
+    serializer_class = BuyerSerializers
+
+
+class SellerViewSet(viewsets.ModelViewSet):
+    queryset = Seller.objects.all()
+    serializer_class = SellerSerializer
+
